refactor(identity): publish auth guards via the ADR-024 contract
Onda 1 (higiene arquitetural): 21 rotas em 5 BCs importavam AuthenticatedUser/authenticated_admin/authenticated_user direto de identity.infrastructure.auth, furando o contrato publicado (ADR-024). Publica os 3 guards no identity/presentation/public.py (a superfície sancionada, ao lado de resolve_profile_id) e migra as 21 rotas para importar de lá. Sem mudança de comportamento — só a fronteira de import; identity fica livre para refatorar o wiring de auth sem rippling nos outros BCs.
