+++ src/modules/media/presentation/routes/admin_credits_routes.py
@@ -14,7 +14,7 @@ from fastapi import APIRouter, Depends
 
 from src.building_blocks.presentation import api_single
 from src.config.containers import ApplicationContainer
-from src.modules.identity.infrastructure.auth import AuthenticatedUser, authenticated_admin
+from src.modules.identity.presentation.public import AuthenticatedUser, authenticated_admin
 from src.modules.media.application.dtos.credits_dtos import (
     ListCreditsStatusInput,
     ResetCreditsDetectionInput,
