+++ src/modules/media/presentation/routes/admin_intro_detection_routes.py
@@ -8,7 +8,7 @@ from fastapi import APIRouter, Depends
 
 from src.building_blocks.presentation import api_list, api_single
 from src.config.containers import ApplicationContainer
-from src.modules.identity.infrastructure.auth import AuthenticatedUser, authenticated_admin
+from src.modules.identity.presentation.public import AuthenticatedUser, authenticated_admin
 from src.modules.media.application.dtos.intro_detection_run_dtos import (
     GetIntroDetectionRunInput,
     ListIntroDetectionRunsInput,
