+++ src/modules/media/presentation/routes/admin_overview_routes.py
@@ -8,7 +8,7 @@ from fastapi import APIRouter, Depends
 
 from src.building_blocks.presentation import api_single
 from src.config.containers import ApplicationContainer
-from src.modules.identity.infrastructure.auth import AuthenticatedUser, authenticated_admin
+from src.modules.identity.presentation.public import AuthenticatedUser, authenticated_admin
 from src.modules.media.application.use_cases.get_library_usage import (
     GetLibraryUsageUseCase,
 )
