+++ src/modules/media/presentation/routes/admin_subtitle_ocr_routes.py
@@ -10,7 +10,7 @@ from fastapi import APIRouter, Depends
 from src.building_blocks.presentation import api_list, api_single
 from src.config.containers import ApplicationContainer
 from src.config.logging import get_logger
-from src.modules.identity.infrastructure.auth import AuthenticatedUser, authenticated_admin
+from src.modules.identity.presentation.public import AuthenticatedUser, authenticated_admin
 from src.modules.media.application.dtos.subtitle_ocr_run_dtos import (
     GetSubtitleOcrRunInput,
     ListSubtitleOcrRunsInput,
