+++ src/modules/media/presentation/routes/tmdb_lookup_routes.py
@@ -18,7 +18,7 @@ from fastapi import APIRouter, Depends, Query
 
 from src.building_blocks.presentation import api_single
 from src.config.containers import ApplicationContainer
-from src.modules.identity.infrastructure.auth import AuthenticatedUser, authenticated_user
+from src.modules.identity.presentation.public import AuthenticatedUser, authenticated_user
 from src.modules.media.application.dtos.tmdb_lookup_dtos import SearchTmdbTitlesInput
 from src.modules.media.application.use_cases.search_tmdb_titles import (
     SearchTmdbTitlesUseCase,
