+++ src/modules/settings/presentation/routes/admin_settings_routes.py
@@ -15,7 +15,7 @@ from fastapi import APIRouter, Depends
 
 from src.building_blocks.presentation import api_list, api_single
 from src.config.containers import ApplicationContainer
-from src.modules.identity.infrastructure.auth import AuthenticatedUser, authenticated_admin
+from src.modules.identity.presentation.public import AuthenticatedUser, authenticated_admin
 from src.modules.settings.application.dtos import UpdateSettingInput
 from src.modules.settings.application.use_cases import (
     ListSettingsUseCase,
